# lxdapi/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.BrandSettings:
    properties:
      admin_bg_image:
        type: string
      admin_bg_opacity:
        type: integer
      admin_content_opacity:
        type: integer
      admin_login_title:
        type: string
      admin_system_name:
        type: string
      admin_system_title:
        type: string
      container_base_template:
        type: string
      container_bg_image:
        type: string
      container_bg_opacity:
        type: integer
      container_content_opacity:
        type: integer
      container_lite_template:
        type: string
      container_login_title:
        type: string
      container_notice:
        type: string
      container_notice_opacity:
        type: integer
      container_system_name:
        type: string
      container_system_title:
        type: string
      created_at:
        type: string
      favicon_url:
        type: string
      footer_text:
        type: string
      id:
        type: integer
      tls_cert_content:
        type: string
      tls_key_content:
        type: string
      updated_at:
        type: string
      use_local_cdn:
        type: boolean
      user_bg_image:
        type: string
      user_bg_opacity:
        type: integer
      user_content_opacity:
        type: integer
      user_login_title:
        type: string
      user_notice:
        type: string
      user_notice_opacity:
        type: integer
      user_system_name:
        type: string
      user_system_title:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/admin/brand-settings:
    get:
      consumes:
      - application/json
      description: 获取系统品牌自定义设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取品牌设置
      tags:
      - Admin API - 品牌设置
    post:
      consumes:
      - application/json
      description: 更新系统品牌自定义设置，reset=true时重置为默认
      parameters:
      - description: 是否重置为默认值
        in: query
        name: reset
        type: string
      - description: 品牌设置
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.BrandSettings'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 更新品牌设置
      tags:
      - Admin API - 品牌设置
  /api/admin/cache/containers:
    get:
      consumes:
      - application/json
      description: 从缓存获取所有容器信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器缓存
      tags:
      - Admin API - 缓存管理
  /api/admin/cache/refresh:
    post:
      consumes:
      - application/json
      description: 刷新容器缓存，无name参数刷新所有，有name参数刷新指定容器
      parameters:
      - description: 容器名称，不传则刷新所有
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 刷新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 刷新缓存
      tags:
      - Admin API - 缓存管理
  /api/admin/captcha:
    get:
      consumes:
      - application/json
      description: 生成图片验证码用于管理员登录
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            properties:
              captcha_id:
                type: string
              code:
                type: integer
              image:
                type: string
            type: object
        "500":
          description: 生成失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      summary: 获取登录验证码
      tags:
      - Admin API - 认证
  /api/admin/containers:
    get:
      consumes:
      - application/json
      description: 获取容器列表，可按用户筛选
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器列表
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name:
    delete:
      consumes:
      - application/json
      description: 删除指定容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 删除容器
      tags:
      - Admin API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器详情
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/action:
    post:
      consumes:
      - application/json
      description: '对容器执行操作: start/stop/restart/pause/resume/reinstall/reset-password/reset-traffic'
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 操作类型
        in: query
        name: action
        required: true
        type: string
      - description: 操作参数
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 容器操作
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/config:
    get:
      consumes:
      - application/json
      description: 获取容器的资源配置信息（仅管理员可用）
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器配置
      tags:
      - Admin API - 容器管理
    put:
      consumes:
      - application/json
      description: 热更新容器配置，支持CPU、内存、磁盘、带宽等升降级，磁盘只能扩容
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 配置更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateContainerConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 更新容器配置
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/credential:
    get:
      consumes:
      - application/json
      description: 获取或创建容器访问凭证
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器凭证
      tags:
      - Admin API - 容器管理
    post:
      consumes:
      - application/json
      description: 重新生成容器访问Hash
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 生成失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 重新生成容器凭证
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/ip:
    get:
      consumes:
      - application/json
      description: 获取容器的IP地址
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 'IP版本: v4/v6/all，默认all'
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器IP
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/ip/allocate:
    post:
      consumes:
      - application/json
      description: 为容器分配IP地址
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      - description: 分配数量
        in: body
        name: request
        schema:
          properties:
            count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 分配失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 分配容器IP
      tags:
      - Admin API - 容器管理
  /api/admin/dashboard:
    get:
      consumes:
      - application/json
      description: 获取系统概览数据，包括容器、用户、系统信息等
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取仪表盘数据
      tags:
      - Admin API - 仪表盘
  /api/admin/host/stats:
    get:
      consumes:
      - application/json
      description: 获取主机CPU、内存、磁盘等实时状态
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取主机状态
      tags:
      - Admin API - 仪表盘
...
swagger: "2.0"